moving terminal velocity and coalescence kernels out of "physics" (as these are not Formulae-handled one-liners). closes #726
