+++ PySDM/attributes/chemistry/acidity.py
@@ -2,7 +2,7 @@
 pH calculated by finding equilibrium hydrogen ion concentration
 """
 from PySDM.attributes.impl.intensive_attribute import DerivedAttribute
-from PySDM.physics.aqueous_chemistry.support import AQUEOUS_COMPOUNDS
+from PySDM.dynamics.impl.chemistry_utils import AQUEOUS_COMPOUNDS
 from PySDM.backends.impl_numba.methods.chemistry_methods import _conc
