+++ PySDM/attributes/impl/mapper.py
@@ -13,7 +13,7 @@ from PySDM.attributes.numerics import CellID, CellOrigin, PositionInCell
 from PySDM.attributes.chemistry import (
     make_mole_amount_factory, make_concentration_factory, Acidity, HydrogenIonConcentration)
 from PySDM.attributes.physics.critical_supersaturation import CriticalSupersaturation
-from PySDM.physics.aqueous_chemistry.support import AQUEOUS_COMPOUNDS
+from PySDM.dynamics.impl.chemistry_utils import AQUEOUS_COMPOUNDS
 from PySDM.physics.surface_tension import Constant
 
 attributes = {
