+++ PySDM/attributes/physics/terminal_velocity.py
@@ -2,7 +2,7 @@
 particle terminal velocity (used for collision probability and particle displacement)
 """
 from PySDM.attributes.impl.derived_attribute import DerivedAttribute
-from PySDM.physics.terminal_velocity.gunn_and_kinzer import Interpolation
+from PySDM.dynamics.terminal_velocity import Interpolation
 
 
 class TerminalVelocity(DerivedAttribute):
