+++ PySDM/dynamics/terminal_velocity/__init__.py
@@ -1,0 +1,4 @@
+"""
+particle terminal velocity formulae
+"""
+from PySDM.dynamics.terminal_velocity.gunn_and_kinzer import Interpolation
