+++ PySDM/physics/__init__.py
@@ -1,5 +1,7 @@
 """
-Physical `PySDM.physics.constants` and formulae
+Physical `PySDM.physics.constants` and simple formulae (essentially one-liners)
+  that can be automatically either njit-ted or translated to C (in contrast to
+  more complex code that resides in backends)
 """
 from . import (condensation_coordinate, latent_heat, saturation_vapour_pressure,
     hygroscopicity, drop_growth, surface_tension, diffusion_kinetics, diffusion_thermics,
