--- PySDM/physics/aqueous_chemistry/__init__.py
@@ -1,4 +1,0 @@
-"""
-support classes and constants for the
- `PySDM.dynamics.aqueous_chemistry.AqueousChemistry` dynamic
-"""
