--- PySDM/physics/terminal_velocity/__init__.py
@@ -1,3 +1,0 @@
-"""
-particle terminal velocity formulae
-"""
