+++ PySDM/products/aqueous_chemistry/aqueous_mass_spectrum.py
@@ -6,7 +6,7 @@ import numpy as np
 from chempy import Substance
 from PySDM.physics.constants import si
 from PySDM.products.impl.spectrum_moment_product import SpectrumMomentProduct
-from PySDM.physics.aqueous_chemistry.support import AQUEOUS_COMPOUNDS
+from PySDM.dynamics.impl.chemistry_utils import AQUEOUS_COMPOUNDS
 
 
 class AqueousMassSpectrum(SpectrumMomentProduct):
